requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-groq
langgraph
chromadb
pypdf
python-dotenv
sentence-transformers 
langchain-huggingface